Formatted for output

+++ Assignment_2/set_1/1.c
@@ -7,16 +7,16 @@ int main()
     float p;
 
     float SI(float, float, int);
-    float CI(float, float, float, int);
+    float CI(float, float, int, int);
 
     scanf("%f", &p);
     
-    printf("Simple Interest for 10 years: %f\n", SI(p, r, t));
-    printf("Compound Interest for 10 years annually: %f\n", CI(p, r, t, 1));
-    printf("Compound Interest for 10 years semi-annually: %f\n", CI(p, r, t, 2));
-    printf("Compound Interest for 10 years quarterly: %f\n", CI(p, r, t, 4));
-    printf("Compound Interest for 10 years monthly: %f\n", CI(p, r, t, 12));
-    printf("Compound Interest for 10 years daily: %f\n", CI(p, r, t, 365));
+    printf("Simple Interest for 10 years: %0.2f\n", SI(p, r, t));
+    printf("Compound Interest for 10 years annually: %0.2f\n", CI(p, r, t, 1));
+    printf("Compound Interest for 10 years semi-annually: %0.2f\n", CI(p, r, t, 2));
+    printf("Compound Interest for 10 years quarterly: %0.2f\n", CI(p, r, t, 4));
+    printf("Compound Interest for 10 years monthly: %0.2f\n", CI(p, r, t, 12));
+    printf("Compound Interest for 10 years daily: %0.2f\n", CI(p, r, t, 365));
 
     return (0);
 }
@@ -27,7 +27,7 @@ float SI(float p, float r, int t)
     return I;
 }
 
-float CI(float p, float r, float t, int n)
+float CI(float p, float r, int t, int n)
 {
     float CI = p * (pow((1 + r/(100*n)), (n * t)) - 1 );
     return CI;
